LPS-188165 Extract logic into a new class, no logic changes
